Make sure the service worker status is always legible
The part of the Service Workers view that provides the current status
is sometimes hidden now.
If you set a huge minimum font size in chrome://settings/fonts, then
the status label gets partly hidden. This is a problem for a11y reasons.

This label conveys information that cannot be gotten otherwise.

This change makes sure the string remains visible and that the panel
adapts to varying available width in those 3 cases:
- normal font
- large zoom
- huge minimum font size setting

Demo: https://imgur.com/a/IUaEHw8

Bug: 1194085
Change-Id: I5aea12765da12ed9e110f4c8de33fdcac7622dba