Allow scrolling within empty tabbed pane placeholders
When a TabbedPane is created, it can optionally receive a placeholder
element that gets displayed when no tabs are shown. For now, only the
Sources panel uses this (to display the ctrl+P shortcut and info about
workspaces).

Other panels also have placeholder-looking areas (network, perf) but
they handle them differently.

The one in Sources does not scroll. So if the area available to display
it is too small, users won't be able to read the full content of the
placeholder text.

This change removes the overflow:hidden which prevented the placeholder
from taking as much space as it needed and therefore prevented the
parent from displaying scrollbars.

Before/after: https://imgur.com/a/aLHLzhy

Bug: 1194486
Change-Id: Iecbf3aa9b34571cedbd5f426b014985a61be5dcf