Update aria-autocomplete and -role attrs for ConsoleView and SuggestBox
This change fixes two accessibility scenarios in the DevTools Console:
- Discovering available auto-complete suggestions
- Reading new console output

The messages container element in ConsoleView was not marked as a live
region so ARIA was not announcing new messages logged to the console,
such as console.log calls, or the result of expressions typed into the
editor.

This change marks the message container element with aria-role="log".
This indicates to a11y tools that the container will occassionally be
updated with new content which should be read aloud to the user. It is
also marked as a "polite" live region as a fallback in case tools don't
recognize this role. New messages are now announced by the screen
reader as they occur.

The SuggestBox class and its consumers have also been updated to
implement the aria-autocomplete attribute following the guidance at
https://w3c.github.io/aria/#aria-autocomplete. I've used the "both"
model for TextPrompt and TextEditorAutocompleteController since these
classes provide both an inline suggestion and a listbox of suggestions.
Screen readers will now recognize the relationship between the text
edit controls and the suggestion list and automatically announce the
first suggestion as the user types.

Bug: 1194704
Change-Id: I958efd333e11aa5077dd0e68f29d9fb8995dd1d5