[DarkMode] cookiesTable
This CL migrates the cookies table away from the legacy colour
patching. However for flagged rows it uses a one off yellow/orange
color. Having chatted with Peter Muller, we decided that we will allow
people to use one off colors if they are defined as variables, and
prefixed with `--override` (which means Stylelint is happy).

The color values here are the exact values that the legacy patching
generated. The removed CSS is a selector that was never used anywhere
in the codebase.

Bug: chromium:1152736
Change-Id: Icf693f768db9097fb4094b8e897049c31ec94e8d