Add lint rule for LitHtml.render calls
This CL adds a rule that ensures that any LitHtml.render call is
nested within a coordinator.write. Note that this rule isn't
completely covering all cases, for example:

```
render() {
  LitHtml.render(...);
}

update() {
  coordinator.write(() => this.render())
}
```

Won't be picked up; but to be honest any case like that feels like one
where we're writing complicated code for the sake of it. This should
pick up most cases.

Note this CL doesn't enable the rule; I'll do that in a follow-up.

Bug: 1188116
Change-Id: Id73eaaa76db4685450086d2f77e7b2de322210d0