Add lint rule for LitHtml.render calls
This CL adds a rule that ensures that any LitHtml.render call is
nested within a coordinator.write. It uses a fairly simple approach of
looking for a parent coordinator.write() call, but it should pick up most
cases.

Note this CL doesn't enable the rule; I'll do that in a follow-up. There
are 22 violations of the rule, so I will work on fixing those.

Bug: 1188116
Change-Id: Id73eaaa76db4685450086d2f77e7b2de322210d0