[Memory] Fix clearing profile view bug
We had a race condition where if the user cleared the profiles in the
Memory panel while recording, we would stil show the profile listing
rather than the launch view. This seems to be because the stopping of
the recording (as part of the clear) dispatches an event to say
profiling has stopped, but when it's handled the profile no longer
exists.

This CL adds a boolean to note when the user cancelled the recording as
a byproduct of clearing the profiles. It uses this flag to early exit
before the event is dispatched.

R=tvanderlippe@chromium.org

Bug: 1191351
Change-Id: I4da0b6f47af9bf32b31bd6875beedffc858497d3