Fix the alignment in console history suggestion
In the console history suggestion, there was a gap
in between the prompt icon and the history suggestion that spread them apart.
This behaviour is used in the Suggestbox of CSS property in the Styles pane (https://crrev.com/c/2485081)
but is not wanted in the Suggestbox in the console because they place their icons on opposite sides.

This CL make it so that there is no gap between the prompt icon and the console history suggestion.

before: https://imgur.com/a/HPNW5YX
after: https://imgur.com/a/pWTtomy

Bug: 1174394
Change-Id: I89d4ecf886d5f05e690ee236043640d7508e6586