Use aria-role="log" for ConsoleView messages container
This is a follow-up change to:
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2798572

The previous changed marked the ConsoleView messages element as an ARIA
live region so that new messages would be announced as they happen.
This change adds the "log" role to this element which better describes
the element's purpose: https://www.w3.org/TR/wai-aria-1.1/#log

Elements marked as "log" are also considered live regions by
accessibility tools but I've kept the markAsPoliteLiveRegion call as-is
for backward compatibility.

The messages element had the "group" role before this change. This was
added so that stack trace messages, which are marked as "treeitem"s,
would have a valid container (must be either "group" or "tree").
However, this meant that all stack trace messages would semantically
belong to a single group.

To ensure stack trace elements still have a valid container after this
change, I've marked each individual stack trace message's
contentElement as a "tree", and moved the aria-expanded property onto
the toggleElement. This maintains the behavior we had before, and
ensures the a11y tree is still valid.

Bug: 1195109
Change-Id: Id6f303eb046c970d1695d181251b38112b5c42de