[css] Properly support reloading SASS source maps via browser-sync.
This aligns SASSSourceMapping to behave similarly to
CompilerScriptMapping, which already supports the case that multiple
sourcemaps provide the same (virtual) source file. With this change we
also get that ability for style sheets, which means that the last added
source map is going to win, and when the latest sourcemap is removed (as
is the case when using browser-sync i.e. via gulp), we fall back to the
previous sourcemap and so on.

Ideally the SASSSourceMapping and CompilerScriptMapping would share most
(if not all) of their code, but that's not immediately possible, so for
the time being they will remain separate and when we get back to the URL
deduplication issue in sourcemaps, we will unify these two classes.

Fixed: chromium:1044383
Change-Id: Id80a2500d99952cc092fa1013106b5f56e4e629c