Change rendering order in the scroll snap overlay
Previously, the container border was rendered before snap areas.
This was leading to the container border being overwritten by
the areas when there are no margins between the container and the
areas. This CL changes the order so that the container border is
rendered on top of the areas.

Before: https://i.imgur.com/iz4kTnM.png
After: https://i.imgur.com/xmpkyKk.png

Fixed: chromium:1196566
Change-Id: Ie3a29013d4c62c72f9f11c55ac12b5e516403bd9