Improve performance of console with many messages
This is the first CL in trying to improve the performance of the
console when slamming it with a lot of messages in quick succession. I
was testing with running:

```
for (let i = 0; i < 1000; i++) console.log(i)
```

The first problem I found is that this causes the sidebar to update
the numbers on the lefthand side. This runs for every single message
and slows down rendering hugely. If I remove the sidebar code, the
console itself updates much more quickly.

I discovered that previously for each new message, for each filter
that the sidebar lists, we constructed a map that contained the
filter's name, and the "compiled" message from i18n(String(string, {
n: messageCount })). For each message, for each filter, we constructed
these strings, but constructed them for all filters, every time we
interated over a single filter. This was therefore taking a lot of
time - 1000 messages, 5 filters, each filter constructing 5 strings,
and then only picking the one that it needs to render.

Tweaking the approach, we now create a Map to lookup the UIString at
runtime, and then only convert that into the fully compiled message
when we update the given filter. This improves the performance by a
good amount because we drastically reduce the calls to i18nString(),
which anecdotally don't seem to be very quick.

There's definitely more to do here - I'd ideally like to update the
sidebar so that it just gets given a count of messages for each filter
- I'm not sure why it has to take every message one at a time and then
figure out the count itself - I think the main `ConsoleView` could
batch calls to, say, `sidebar.updateMessageCounts(...)`, and only run
that (say) every 100ms. That would keep the UI much snappier and I
don't think it's a disaster if the sidebar is momentarily showing
outdated counts.

The initial bug report also mentions event listeners as a potential
performance issue. I've yet to dive into that as the UIString issue
outlined above was the first thing that caught my eye, but I will
continue to investigate.

Bug: 1195494
Change-Id: Ib399f03e7aa1a3298762c1fb755936dd6586e1cf