Fix ComponentServer relying on <style> tag to inject deps
The component server needs to inject CSS and JS scripts into each
example. It was doing this by relying on a <style> and <script> tag as
the hook to inject more code. But if you have an example without a
<style> or <script> tag, it won't work. Instead we now inject based on
the </head> and </body> tags, which will always exist (or, if they
don't, we have bigger problems!)

Bug: None
Change-Id: Id2f586c917ee7ab0821d5ab8e50e5cd3d3647a92