Don't update console sidebar if it's not visible
The console sidebar is slow to update if we have many messages at once
and will slow the updating of the console by seconds if given a lot of
messages at once.

The worst thing is that users paid this cost whether they have the
sidebar open or not. This CL updates the code to queue messages should
the sidebar be closed, and then fire them once it opens. This means if
the sidebar is closed, the console is much snappier. This CL does
nothing to improve the perf should the sidebar be open, but we can do
more for that later.

Bug: 1195494
Change-Id: Ica886fa730de8934d0c742921b4d1611b5a0daa4