Skip over overridden properties when looking for a css variable to jump to
The current logic to jump from a var() function to where the variable is
declared is to iterate over all rules and all properties from top to
bottom and stop at the first matching variable name.
This works well in most cases as properties in higher specificy rules
appear above properties in lower specificy rules.
It fails in the case where a property is defined twice in the same rule.
In this case the last one wins based on source order.

So we now skip over overridden properties.

Demo: https://imgur.com/eoSda9Y

Bug: 1192801
Change-Id: Ie3a55531b34b4b83775a9aaf4e35c7d0d1507d28