Handle duplicate track property changes
Adds a quick hash function for the AttributeList so that they can be
compared without checking content each time. This is used to prevent
recreating, removing, and re-adding the same node (which would break
user focus on a particular tab).

Its only recently that tracks can change content in the middle of a
video, which also uncovered a typo where _textTracksTabs was being used
but the actual object was being stored in _textTrackTabs, so that's
fixed too.

R=tvanderlippe@chromium.org

Change-Id: I0858071960ce77a0be1530df18fb463932dc4d8b