Prevent the close button from ever disappearing
Many people open devtools by accident, then try to get rid of it by
making it smaller to the side, and then complain that they can't find
the close button anymore.

This change prevents the close button from ever getting out of the
viewport when space is missing.
This is only important when docked since undocked mode has the close
button as part of the window.

We should also make other changes to avoid other buttons from
disappearing, but this is a much needed change for a common problem.

Before/after: https://imgur.com/a/vhgmHhM

Bug: 1171131
Change-Id: I28ca0e8ecea6ad0e6660a733190d9b37ab424619