[A11yTree] [WIP] wire up DOM selected Node and A11y tree
Do Not Submit

This allows the user to inspect a node in the DOM, and then flick
over to the A11y tree and see the same node selected/highlighted. This
can involve a round trip to the backend to fetch more of the tree,
so introduces a new method into AccessibilityModel:

- requestAndLoadSubTreeToNode
  Given a DOMNode, get the subtree containing parent/ommer nodes.
  Currently this does not re-request the node if it has already been
  fetched. A bug or WAI? Only time will tell.

There are some kinks though. Currently, when we get a node back from
this method and use it to expand the tree, I am re-creating the
TreeNode<..> that wraps it, the base unit the TreeOutline component
uses. This causes issues with strict equality, so this CL contains a
gross hack to fix that, which needs to be worked on with jacktfranklin@.

It also does not highlight the node when you do "right click >
Inspect", although the tree will be loaded correctly. This is
presumably due to the fact that nothing sets initial focus when the
A11y tree is first seen, either on the root node or a specific node.

TBD

Bug: 887173

Change-Id: I77aff0a391b0307e4f9a53ba6416fe4a027be84c