Add UA metadata to Network Conditions panel
Part I of a set of changes to improve the experience for UA Client Hints
in the Network Conditons panel.

This change adds a 'metadata' property with an appropriate
UserAgentMetadata object to each pre-canned user agent object listed in
NetworkConfigView.ts.

User agents that do not support UA Client Hints have their metadata set
to null to indicate that navigator.userAgentData should not be
overridden.

A follow-up change will actually use this metadata to override the
metadata seen by the page when a pre-canned user agent is chosen.

Bug: 1174299
Change-Id: If41cf7f4e8fa074ff28bb7bd762f7a5ca443bae0