[TreeOutline] Add optional ID param to TreeNodes
To support use cases where strict object equality is not feasible, this
CL adds in an optional id parameter that can be specified per node. The
components traversal algorithms and focus behaviour will then rely on
this id, rather than object equality.

Devised specifically for the AccessibilityTree, this is mostly a not
very robust hack. The id type is currently hard coded as a string,
but possibly should be a generic key type that can be specified by the
component author.

Also, instead of styling on :focus, the TreeOutline component now keeps
track of the selected node, and styles based on that. The benefit being
that when focus moves away from the tree (or devtools), the selected
node is still highlighted.

Bug: 887173, 1166669
Change-Id: If493a7c7fdf1d9d44fffda7940fb96e4438b9e53