[A11yTree] wire up DOM selected Node and A11y tree
Do Not Submit

This allows the user to inspect a node in the DOM, and then flick
over to the A11y tree and see the same node selected/highlighted. This
can involve a round trip to the backend to fetch more of the tree,
so introduces a new method into AccessibilityModel:

- requestAndLoadSubTreeToNode
  Given a DOMNode, get the subtree containing parent/ommer nodes.
  Currently this does not re-request the node if it has already been
  fetched. A bug or WAI? Only time will tell.

This relies on there not being strict object equality in the TreeOutline
so is dependant on https://crrev.com/c/2827554.

Future improvements:
- account for Ignored nodes in the tree. Eg: inspecting <head> or
  something that is aria-hidden. This change adds in the rendering
  for Ignored nodes, but not the logic to handle them in the tree.
  They will need to be dynamically inserted and removed in the tree,
  as they should appear only when inspected, but should not be able
  to be navigated to/explored

Bug: 887173

Change-Id: I77aff0a391b0307e4f9a53ba6416fe4a027be84c