Make incremental rebuilds work for ts_library()
Prior to the resources/inspector removal, we were having efficient
recompilations with GN. However, after the resources/inspector removal,
GN was excessively recompiling targets that don't require recompilation.

After an extensive investigation, we discovered that the root cause is
a discrepancy between what `tsc` determines to be a "new" file and when
Ninja does. Essentially, `tsc` uses content-based hashing to determine
freshness, whereas Ninja looks at file timestamps.

Therefore, if `tsc` generated the same output, `tsc` thought that nothing
had changed. However, `tsc` always writes the output, meaning that the
filestamps changed and then Ninja picked that up concluding "I need to
rerun dependent actions, as the output changed".

Therefore, we need to mimic what Ninja expects based on the file contents
of what `tsc` generates. To do so, we augment `ts_library` to obtain
any previously generated file content and store the corresponding file
timestamps. After `tsc` did its job, we compare the newly generated
content to the previous content. If we determine that they are equivalent,
we force-reset the timestamps of the files. This allows Ninja to conclude
that nothing has changed.

We do a similar fix for the generated `tsconfig.json`, but there we can
directly compare the newly generated configuration with the old configuration
and only perform the write if necessary.

The reason this improves incremental compilation is the fact that we split
up the implementation of a module with the entrypoint of a module. If you
make an implementation change in a module, the `ts_library` will have changed,
which means that it directs dependents will run. Since the only direct
dependent of a `devtools_module` is the corresponding `devtools_entrypoint`,
the `devtools_entrypoint` action will run. However, since the
`devtools_entrypoint` action will generate an equivalent public API, all
users of the entrypoint will not rebuild. Therefore, implementation-only
changes will now only cause recompilations of the module, but none of their
dependents.

R=aerotwist@chromium.org
CC=dpranke@google.com

Also-By: dpranke@google.com
Bug: 1139220
Change-Id: I9cafe33f92ce70fb638884b850c873987687d5a4