Correctly resolve relative sourceURL annotations.
As a backwards compatible first fix, we are trying to resolve the
sourceURL annotation in the DevTools front-end relative to the
embedderName (the actual URL in case of Chromium). We need to remember
the url that V8 reported and use that when setting a breakpoint though.

This is still not ideal, since the front-end doesn't really know where
the script is comining from, but it's a pragmatic, backwards-compatible
first step.

Before: https://imgur.com/RshpZa5.png
After: https://imgur.com/wBrlStz.png
Bug: chromium:1183990, b/180404336
Change-Id: I762b998ae90e168665967741c7fd351ec2ceb905