Add issues for CORS invalid credentialed requests
This CL adds issues for the CORS error codes

InvalidAllowCredentials
PreflightInvalidAllowCredentials

This also adds the $host_port replacement for headers in .rawresponse
files. This is useful if a HTTP header need to whitelist an origin
(which includes the port). Since our test setup changes the port
we host on every time, this placehoder is used in
test/e2e/resources/issues/acac-invalid.rawresponse

Screenshot: https://imgur.com/a/6lhV7WB

Bug: chromium:1141824
Change-Id: I9fb4a944241a5479b55cf5f97d46001346bb8a26