[BUILD.gn] Assert grd files when building chrome target
This adds a dependency on assert_grd when builiding the chrome
target. Our CQ bots are only building the chrome target, and with the
previous dependency chain, the existence of grd files were not checked.
This CL adds a dependency on the assert_grd action, which makes sure
that we correctly generate the files when passing CQ.

Also-by: tvanderlippe@chromium.org
Bug: none
Change-Id: I2beeee62dcf4adc3881a5a93d8751448b667dc1c