Cleanup front_end/BUILD.gn
We only need to list the entrypoints and any special actions in
the front_end/BUILD.gn. Any other action is transitively depended
on by the relevant entrypoints.

This means that if a folder becomes unused by the entrypoints,
Ninja will now start to complain about files being missed (rightfully
so). Therefore, we allow us to prune the GRD file whenever possible.

R=alexrudenko@chromium.org

Bug: 1174013
Change-Id: I5473c7f6f39f5fb92592692c67caf96f805a2a22