Update theme colors component docs
This CL updates the theme colors doc so that the actual color values
come from CSS, so they don't get outdated. Additionally we now render
into a div that has the .-theme-with-dark-background class so that the
dark colors are rendered correctly too. Long term I'd like to generate
the entire list of variables automatically,but this is already an
improvement.

Bug: chromium:1152736
Change-Id: I89e44a9060fbef0866f670cced2030fdb2443526