Modify a couple large SVG icons to avoid losing them when SVGO optimizes
This might have started with a recent SVGO change, but the fact is that
2 of the icons from front_end/Images/src/largeIcons.svg are now missing
in the optimized front_end/Images/largeIcons.svg file.

It looks like the styleToAttr SVGO plugin removes the fill on these 2
icons, and we end up with transparent icons.
Just moving the fill from the individual <path> elements to the <g>
group fixes that issue.

The other option would have been to add an SVGO config file to disable
the styleToAttr plugin, but that requires adding a javascript file to
contain the configuration and also requires installing SVGO as a
dependency (whereas we don't do it now, thanks to npx).

Optimized SVG before/after: https://imgur.com/a/sIe1G0h

Bug: 1203248
Change-Id: Id18c55987e657778efd4da15485977d9e6eae849