Make autocomplete icons context-aware to support CSS grid icons
Before we can start working on a CSS grid editor, we should
support CSS Grid specific behaviour for autocomplete icons
in the Styles panel. This CL renames some functions to
make them generic enough to be used for CSS Grid icons,
fixes an issue with parent computed styles being wrong and
makes the icons to be dependent on the computed display
property of the container element.

Bug: chromium:1203241
Change-Id: I39e37deec011119d8a8c062186230940ad8c5959