Use an AdornerRegistry for adorner names and categories
Previously we create literal strings directly at adorner's create site
to assign adorner texts, as well as when creating options in the adorner
control pane. These strings have a potential to mismatch, and in general
are prone to bugs. This CL changes this into a predefined
AdornerDefinition, so that clients using any adorners won't misspell
adorner names or misuse adorner categories. It also forces clients to
be explicit about adorner categories.

Bug: chromium:1066772
Change-Id: I0753a73a590260ba26220254cc4d61232a9c704e