Fix FrameManager's handling of OOPIF transitions
When a frame is being turned into an OOPIF, the frame is being detached
from its original target and added to a new target. Usually the order of
events is:
1) added to original target
2) added to new target
3) detached from original target

But there is no guarantee for this order of events and sometimes the
order of steps 2) and 3) is inverted (detected by a flaky test on the
Mac bot). This CL ensures that the FrameManager is able to correctly
handle both cases. This mainly means taking care of copying frame
creation stack traces.

Disable test CL: https://crrev.com/c/2862803
Update and enable test CL: https://crrev.com/c/2862770

Bug: 1202024
Change-Id: Ie42efcaf75c4aa8a0e71649c9790b23c7eff9e33