[A11yTree] Handle ignored inspected nodes
Nodes that are ignored/uninteresting for accessibility are exposed in
the devtools accessibility tree using a placeholder, that shows
the "Ignored" status. These nodes are not able to be navigated to by
the tree UI itself as they are excluded from the tree that is built via
the backend, except in the case where the currently inspected DOM node
corresponds to one of these Ignored nodes. In this case, a transient
node is created and selected so that the user can view information about
why this node was Ignored in the ARIA-attributes sub pane, but it is
destroyed and removed once the user navigates away or inspects a new node.

The fix for this involved implementing an extra layer of caching around
the TreeOutline component, so the AccessibilityTreeView controller could
add/remove nodes as needed. In this case the TreeOutline consumer needs
to handle caching, an optional parameter was also added into the
TreeOutlineData struct to disable the internal caching layer, as this
made handling ignored nodes impossible in this case. It's a hack, but
it works.

Change-Id: Ia8c41711a31c307775f8053b0e7a9347b671e91b

Bug: 887173
Change-Id: Ia8c41711a31c307775f8053b0e7a9347b671e91b