Changed height to min-height in media-property-renderer class in playerPropertiesView.css to prevent overlapping of text when wrapped
Height is changed to min-height in media-property-renderer class in playerPropertiesView.css
to prevent overlapping of text when the player properties section's width is reduced.
The height property gives the media-property-renderer section a fixed height of 28px
but when the text content inside it is wrapped its height exceeds 28px
causing text to overflow and appear overlapping over the text beneath.

Changing it to min-height would cause the container to adjust height according to content.

Before: https://imgur.com/OlXv05z

After: https://imgur.com/QopZg0r

Bug: 1205347
Change-Id: I50566e5037adb757bf3491f011fd03448f5de4fa