Adds support for statics in templates
Historically we couldn't do <${tag}> in the DevTools frontend. With the
arrival of Lit2 we can now do this, but there is potential for
performance impact as the implementation of Lit2's statics means that it
could recompute the template on every render.

While we figure out what the best long-term solution is, this CL adds
a wrapper for the Lit html function that supports our own (more limited)
static handler that provides the ability to do <${tag}>, but at the
expense of not being able to update the static values after the fact.

Follow-up: move all call sites to use the indirect form of Lit exported
here, and remove the visibility of the third_party Lit.

R=jacktfranklin@chromium.org

Bug: 1207302
Change-Id: I183fdf93f90950c8bf0a0eee6a5bcc9c5a539a48